ESP8266 and passive mode fixes
ESP8266 appears to always return `1` from `UART.any()`

When in passive mode, the sensor will only reply with a frame if a `0xe2` command is sent. If the sensor is already in passive mode and the controller restarts, `wakeUp` would not  request data and therefore never mark the device as alive.